+++ Sources/KeyboardKit/Appearance/Image+Button.swift
@@ -52,7 +52,6 @@ private extension Image {
     }
 }
 
-
 struct ImageButton_Previews: PreviewProvider {
     
     static var images: [Image] = [